+++ R/giotto_structures.R
@@ -842,7 +842,6 @@ spline_poly <- function(xy, vertices = 20, k = 3, ...) {
 #' @param vertices number of vertices
 #' @param k k
 #' @param set_neg_to_zero set negative values to zero (default: TRUE)
-#' @param verbose be verbose (default: TRUE)
 #' @param ... additional params to pass to \code{spline}
 #' @return Smoothed Giotto polygon object with reduced vertices
 #' @concept polygon
@@ -901,17 +900,15 @@ smoothGiottoPolygons = function(gpolygon,
 
   new_spatvec = dt_to_spatVector_polygon(comb_res)
 
-  if(isTRUE(verbose)) {
-    for(ID in new_spatvec$poly_ID) {
-      bool = terra::is.valid(new_spatvec[new_spatvec$poly_ID == ID])
-      if(!isTRUE(bool)) {
-        print(ID)
-        #plot(new_spatvec[new_spatvec$poly_ID == ID])
-        #orig_spatvector = gpolygon@spatVector
-        #new_spatvec[new_spatvec$poly_ID == ID] = orig_spatvector[orig_spatvector$poly_ID == ID]
-      }
-    }
-  }
+  # for(ID in new_spatvec$poly_ID) {
+  #   bool = terra::is.valid(new_spatvec[new_spatvec$poly_ID == ID])
+  #   if(!isTRUE(bool)) {
+  #     print(ID)
+  #     #plot(new_spatvec[new_spatvec$poly_ID == ID])
+  #     #orig_spatvector = gpolygon@spatVector
+  #     #new_spatvec[new_spatvec$poly_ID == ID] = orig_spatvector[orig_spatvector$poly_ID == ID]
+  #   }
+  # }
 
   new_gpolygon = create_giotto_polygon_object(name = gpolygon@name,
                                               spatVector = new_spatvec,