+++ NEWS.md
@@ -1,5 +1,8 @@
 # TBD RELEASE
 
+## Breaking Changes
+- Removed support for deprecated nesting in `@nn_network` slot
+
 ## Added
 - New `giottoToAnnData()` interoperability function to convert giotto object to squidpy flavor anndata .h5ad file(s)
 - Add missing `create_nn_net_obj()` internal constructor function for S4 `nnNetObj`